requirements.txt
flask>=3.0.0
groq>=0.9.0
faiss-cpu>=1.7.4
sentence-transformers>=2.6.0
transformers>=4.38.0
torch>=2.0.0
deep-translator>=1.11.4
langdetect>=1.0.9
pdfplumber>=0.10.3
numpy>=1.24.0
